Make threadpool module a required dependency
Make the threadpool module a required dependency since it
doesn't pull in any other dependencies and is rather light
weight. This is done since it seems that optional library
dependencies are an antipattern [1]

[1] - http://doc.crates.io/manifest.html#usage-in-packages

Closes #27
