+++ Cargo.toml
@@ -12,11 +12,7 @@ keywords = ["statsd", "metrics"]
 
 [dependencies]
 log = "0.3"
-threadpool = { version = "1.3.1", optional = true }
-
-[features]
-default = ["threading"]
-threading = ["threadpool"]
+threadpool = "1.3.1"
 
 [lib]
 name = "cadence"
