+++ src/sinks/mod.rs
@@ -10,7 +10,6 @@
 
 
 use log::LogLevel;
-use threadpool::ThreadPool;
 
 use std::io;
 use std::io::Write;
@@ -20,6 +19,8 @@ use std::sync::{Arc, Mutex};
 use ::io::{MultiLineWriter, UdpWriteAdapter};
 use ::types::{MetricResult, MetricError, ErrorKind};
 
+pub mod threading;
+
 
 // Default size of the buffer for buffered metric sinks. This
 // is a rather conservative value, picked to make sure the entire
@@ -27,14 +28,6 @@ use ::types::{MetricResult, MetricError, ErrorKind};
 // different value based on the configuration of the network
 // their application runs in.
 const DEFAULT_BUFFER_SIZE: usize = 512;
-
-
-// Default size (number of threads) for the thread pool used by
-// the `AsyncMetricSink` for sending metrics. This value is rather
-// arbitrary but should be fine for most use cases. Users that
-// need further customization can use the alternate constructor
-// for the `AsyncMetricSink`.
-const DEFAULT_THREAD_POOL_SIZE: usize = 4;
 
 
 /// Trait for various backends that send Statsd metrics somewhere.
@@ -292,124 +285,6 @@ impl MetricSink for BufferedUdpMetricSink {
 }
 
 
-/// Implementation of a `MetricSink` that wraps another implementation
-/// and uses it to emit metrics asynchronously with a thread pool.
-///
-/// The wrapped implementation can by any thread safe (`Send + Sync`)
-/// `MetricSink` implementation. Results from the wrapped implementation
-/// will be discarded.
-///
-/// Because this `MetricSink` implementation uses a thread pool, the sink
-/// itself cannot be shared between threads. Instead, callers may opt to
-/// create a `.clone()` for each thread that needs to emit metrics. This
-/// of course requires that the wrapped sink implements the `Clone` trait
-/// (all of the sinks that are part of Cadence implement `Clone`).
-///
-/// When cloned, the new instance of this sink will have a cloned thread
-/// pool instance that submits jobs to the same worker threads as the
-/// original sink and a reference (`Arc`) to the wrapped sink. If you
-/// plan on cloning this sink, the thread pool should be sized
-/// appropriately to be used by all expected sink instances.
-#[derive(Debug, Clone)]
-pub struct AsyncMetricSink<T: 'static + MetricSink + Send + Sync> {
-    pool: ThreadPool,
-    delegate: Arc<T>,
-}
-
-
-impl<T: 'static + MetricSink + Send + Sync> AsyncMetricSink<T> {
-    /// Construct a new `AsyncMetricSink` instance wrapping another sink
-    /// implementation.
-    ///
-    /// The `.emit()` method of the wrapped sink will be executed in a
-    /// different thread via a thread pool. The wrapped sink should be
-    /// thread safe (`Send + Sync`).
-    ///
-    /// The default thread pool size is four threads. Callers can use
-    /// more or fewer threads by making use of the `with_threadpool`
-    /// constructor.
-    ///
-    /// # UDP Sink Example
-    ///
-    /// In this example we wrap the basic UDP sink to execute it in a
-    /// different thread.
-    ///
-    /// ```no_run
-    /// use std::net::UdpSocket;
-    /// use cadence::{UdpMetricSink, AsyncMetricSink, DEFAULT_PORT};
-    ///
-    /// let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-    /// let host = ("metrics.example.com", DEFAULT_PORT);
-    /// let udp_sink = UdpMetricSink::from(host, socket).unwrap();
-    /// let async_sink = AsyncMetricSink::from(udp_sink);
-    /// ```
-    ///
-    /// # Buffered UDP Sink Example
-    ///
-    /// This example uses the buffered UDP sink, wrapped in the async
-    /// metric sink.
-    ///
-    /// ```no_run
-    /// use std::net::UdpSocket;
-    /// use cadence::{BufferedUdpMetricSink, AsyncMetricSink, DEFAULT_PORT};
-    ///
-    /// let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-    /// let host = ("metrics.example.com", DEFAULT_PORT);
-    /// let udp_sink = BufferedUdpMetricSink::from(host, socket).unwrap();
-    /// let async_sink = AsyncMetricSink::from(udp_sink);
-    /// ```
-    pub fn from(sink: T) -> AsyncMetricSink<T> {
-        Self::with_threadpool(sink, ThreadPool::new(DEFAULT_THREAD_POOL_SIZE))
-    }
-
-    /// Construct a new `AsyncMetricSink` instance wrapping another sink
-    /// implementation that will use the provided thread pool for
-    /// asynchronous execution.
-    ///
-    /// The `.emit()` method of the wrapped sink will be executed in a
-    /// different thread via a thread pool. The wrapped sink should be
-    /// thread safe (`Send + Sync`).
-    ///
-    /// # Buffered UDP Sink With a Single Threaded Pool
-    ///
-    /// ```no_run
-    /// # extern crate threadpool;
-    /// # extern crate cadence;
-    /// # fn main() {
-    /// use std::net::UdpSocket;
-    /// use threadpool::ThreadPool;
-    /// use cadence::{BufferedUdpMetricSink, AsyncMetricSink, DEFAULT_PORT};
-    ///
-    /// let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-    /// let host = ("metrics.example.com", DEFAULT_PORT);
-    /// let udp_sink = BufferedUdpMetricSink::from(host, socket).unwrap();
-    /// let async_sink = AsyncMetricSink::with_threadpool(
-    ///     udp_sink, ThreadPool::new(1));
-    /// # }
-    /// ```
-    pub fn with_threadpool(sink: T, pool: ThreadPool) -> AsyncMetricSink<T> {
-        AsyncMetricSink {
-            pool: pool,
-            delegate: Arc::new(sink),
-        }
-    }
-}
-
-
-impl<T: 'static + MetricSink + Send + Sync> MetricSink for AsyncMetricSink<T> {
-    fn emit(&self, metric: &str) -> io::Result<usize> {
-        let owned_metric = metric.to_string();
-        let sink = self.delegate.clone();
-
-        self.pool.execute(move || {
-            let _r = sink.emit(&owned_metric);
-        });
-
-        Ok(metric.len())
-    }
-}
-
-
 /// Implementation of a `MetricSink` that discards all metrics.
 ///
 /// Useful for disabling metric collection or unit tests.
@@ -473,8 +348,7 @@ mod tests {
     use std::net::UdpSocket;
 
     use super::{get_addr, MetricSink, NopMetricSink, ConsoleMetricSink,
-                LoggingMetricSink, UdpMetricSink, BufferedUdpMetricSink,
-                AsyncMetricSink};
+                LoggingMetricSink, UdpMetricSink, BufferedUdpMetricSink};
 
     #[test]
     fn test_get_addr_bad_address() {
@@ -539,12 +413,5 @@ mod tests {
         // the end.
         assert_eq!(9, sink.emit("foo:54|c").unwrap());
         assert_eq!(9, sink.emit("foo:67|c").unwrap());
-    }
-
-    #[test]
-    fn test_async_nop_metric_sink() {
-        let sink = AsyncMetricSink::from(NopMetricSink);
-        assert_eq!(8, sink.emit("buz:33|c").unwrap());
-        assert_eq!(8, sink.emit("boo:27|c").unwrap());
     }
 }
